@@ -2,22 +2,21 @@
     "name": "justbetter/magento2-akeneo-bundle",
     "description": "Magento2 bundle for extending the Akeneo connector with awesome features.",
     "require": {
-        "php": ">=8.0",
+        "php": "^8.2 || ^8.3 || ^8.4",
         "akeneo/module-magento2-connector-community": "*"
     },
     "require-dev": {
         "bitexpert/phpstan-magento": "^0.11.0",
-        "phpstan/phpstan": "^1.10"
+        "phpstan/phpstan": "^1.10",
+        "phpro/grumphp": "^2.0",
+        "magento/magento-coding-standard": "^33",
+        "squizlabs/php_codesniffer": "^3.0"
     },
     "type": "magento2-module",
     "authors": [
         {
-            "name": "Pim Ruiter",
-            "email": "pim@justbetter.nl"
-        },
-        {
-            "name": "Robin Mulder",
-            "email": "robin@justbetter.nl"
+            "name": "JustBetter B.V.",
+            "email": "hello@justbetter.nl"
         }
     ],
     "repositories": {
@@ -37,7 +36,9 @@
     "config": {
         "allow-plugins": {
             "php-http/discovery": true,
-            "magento/composer-dependency-version-audit-plugin": true
+            "magento/composer-dependency-version-audit-plugin": true,
+            "dealerdirect/phpcodesniffer-composer-installer": true,
+            "phpro/grumphp": true
         }
     }
 }
