@@ -6,10 +6,8 @@ use Akeneo\Connector\Helper\Import\Product as BaseProduct;
 
 class Product extends BaseProduct
 {
-    /** Overwritten to set values for store view 0 when attributes are scopable/localizable and required */
     protected function getColumnsFromResult(array $result, array $keys = []): array
     {
-        // This returns the result for the temp table DB columns. ex: 'name-nl_NL-ecommerce' => 'value'
         $mappedResult = parent::getColumnsFromResult($result, $keys);
 
         if (!$this->scopeConfig->getValue('akeneo_connector/justbetter/defaultstorevalues') || !array_key_exists('values', $result)) {
@@ -17,23 +15,19 @@ class Product extends BaseProduct
         }
 
         $adminChannel = $this->scopeConfig->getValue('akeneo_connector/akeneo_api/admin_channel');
-
         $defaultLanguage = $this->scopeConfig->getValue('akeneo_connector/justbetter/defaultlanguage');
-
         $requiredAttributes = $this->getRequiredAttributes();
 
         foreach ($requiredAttributes as $requiredAttribute) {
-
-            if (
-                !array_key_exists($requiredAttribute, $result['values']) ||
-                count($result['values'][$requiredAttribute]) == 0 ||
+            if (!array_key_exists($requiredAttribute, $result['values']) ||
+                count($result['values'][$requiredAttribute]) === 0 ||
                 !$this->isScopableOrLocalizable($requiredAttribute, $mappedResult)
             ) {
                 continue;
             }
 
-            if (!array_key_exists($requiredAttribute.'-'.$defaultLanguage.'-'.$adminChannel, $mappedResult) && $defaultLanguage) {
-                $mappedResult[$requiredAttribute.'-'.$defaultLanguage.'-'.$adminChannel] = $this->getFirstValue($result['values'][$requiredAttribute]);
+            if (!array_key_exists($requiredAttribute . '-' . $defaultLanguage . '-' . $adminChannel, $mappedResult) && $defaultLanguage) {
+                $mappedResult[$requiredAttribute . '-' . $defaultLanguage . '-' . $adminChannel] = $this->getFirstValue($result['values'][$requiredAttribute]);
             }
 
             $mappedResult[$requiredAttribute] = $this->getFirstValue($result['values'][$requiredAttribute]);
@@ -45,21 +39,20 @@ class Product extends BaseProduct
     protected function getFirstValue(array $values): mixed
     {
         $array = array_reverse($values);
+
         return array_pop($array)['data'] ?? '';
     }
 
-    /** Check if an attribute is scopeable or localizable based on the column result name, ex. name-nl_NL-ecommerce */
     protected function isScopableOrLocalizable(string $attributeCode, array $columnResult): bool
     {
         $columns = array_keys($columnResult);
 
         foreach ($columns as $column) {
-
             if ($column === $attributeCode) {
                 return false;
             }
 
-            if (substr($column, 0, strlen($attributeCode)) === $attributeCode) {
+            if (str_starts_with($column, $attributeCode)) {
                 return true;
             }
         }
@@ -67,7 +60,6 @@ class Product extends BaseProduct
         return false;
     }
 
-    /** Get a list of Magento's required product attributes */
     protected function getRequiredAttributes(): array
     {
         $eavAttributeTable = $this->connection->getTableName('eav_attribute');
@@ -75,8 +67,10 @@ class Product extends BaseProduct
 
         $select = $this->connection->select()
             ->from("$eavAttributeTable AS attr")
-            ->join("$eavEntityTypeTable AS type",
-                "attr.entity_type_id = type.entity_type_id AND type.entity_type_code = 'catalog_product'")
+            ->join(
+                "$eavEntityTypeTable AS type",
+                "attr.entity_type_id = type.entity_type_id AND type.entity_type_code = 'catalog_product'"
+            )
             ->where('is_required = 1');
 
         $requiredAttributes = $this->connection->fetchAll($select);
