@@ -1,24 +1,27 @@
 <?php
 
+declare(strict_types=1);
+
 namespace JustBetter\AkeneoBundle\Block\Adminhtml\Akeneo\Edit;
 
 use Magento\Backend\Block\Widget\Form\Generic;
 
 class Form extends Generic
 {
-    /**
-     * Prepare form
-     *
-     * @return $this
-     */
-    protected function _prepareForm()
+    protected function _prepareForm(): Form
     {
-        /** @var \Magento\Framework\Data\Form $form */
-        $form = $this->_formFactory->create(
-            ['data' => ['id' => 'edit_form', 'action' => $this->getData('action'), 'method' => 'post' , 'enctype' => 'multipart/form-data']]
-        );
+        $form = $this->_formFactory->create([
+            'data' => [
+                'id' => 'edit_form',
+                'action' => $this->getData('action'),
+                'method' => 'post',
+                'enctype' => 'multipart/form-data'
+            ]
+        ]);
+
         $form->setUseContainer(true);
         $this->setForm($form);
+
         return parent::_prepareForm();
     }
 }
