@@ -1,59 +1,30 @@
 <?php
 
+declare(strict_types=1);
+
 namespace JustBetter\AkeneoBundle\Block\Adminhtml\Akeneo;
 
+use JustBetter\AkeneoBundle\Model\ResourceModel\Akeneo\CollectionFactory;
+use JustBetter\AkeneoBundle\Model\Status;
+use Magento\Backend\Block\Template\Context;
+use Magento\Backend\Block\Widget\Grid\Extended;
 use Magento\Backend\Helper\Data;
 use Magento\Framework\Module\Manager;
-use Magento\Backend\Block\Template\Context;
-use JustBetter\AkeneoBundle\Model\Status;
-use Magento\Backend\Block\Widget\Grid\Extended;
-use JustBetter\AkeneoBundle\Model\AkeneoFactory;
-use JustBetter\AkeneoBundle\Block\Adminhtml\Akeneo\Grid as GridOption;
 
 class Grid extends Extended
 {
-    /**
-     * @var Manager
-     */
-    protected $moduleManager;
-
-    /**
-     * @var akeneoFactory
-     */
-    protected $_akeneoFactory;
-
-    /**
-     * @var Status
-     */
-    protected $_status;
-
-    /**
-     * @param Context         $context
-     * @param Data            $backendHelper
-     * @param AkeneoFactory   $AkeneoFactory
-     * @param Status          $status
-     * @param Manager         $moduleManager
-     * @param array           $data
-     *
-     */
     public function __construct(
         Context $context,
         Data $backendHelper,
-        AkeneoFactory $AkeneoFactory,
-        Status $status,
-        Manager $moduleManager,
+        protected CollectionFactory $collectionFactory,
+        protected Status $status,
+        protected Manager $moduleManager,
         array $data = []
     ) {
-        $this->_akeneoFactory = $AkeneoFactory;
-        $this->_status = $status;
-        $this->moduleManager = $moduleManager;
         parent::__construct($context, $backendHelper, $data);
     }
 
-    /**
-     * @return void
-     */
-    protected function _construct()
+    protected function _construct(): void
     {
         parent::_construct();
         $this->setId('postGrid');
@@ -64,13 +35,9 @@ class Grid extends Extended
         $this->setVarNameFilter('post_filter');
     }
 
-    /**
-     * @return $this
-     */
-    protected function _prepareCollection()
+    protected function _prepareCollection(): Grid
     {
-        /* @phpstan-ignore-next-line */
-        $collection = $this->_akeneoFactory->create()->getCollection();
+        $collection = $this->collectionFactory->create();
         $this->setCollection($collection);
 
         parent::_prepareCollection();
@@ -78,18 +45,14 @@ class Grid extends Extended
         return $this;
     }
 
-    /**
-     * @return $this
-     * @SuppressWarnings(PHPMD.ExcessiveMethodLength)
-     */
-    protected function _prepareColumns()
+    protected function _prepareColumns(): Grid
     {
         $this->addColumn(
             'id',
             [
-                'header'           => __('ID'),
-                'type'             => 'number',
-                'index'            => 'id',
+                'header' => __('ID'),
+                'type' => 'number',
+                'index' => 'id',
                 'header_css_class' => 'col-id',
                 'column_css_class' => 'col-id',
             ]
@@ -98,10 +61,10 @@ class Grid extends Extended
         $this->addColumn(
             'import',
             [
-                'header'  => __('Type'),
-                'index'   => 'import',
-                'type'    => 'options',
-                'options' => GridOption::getOptionArray0(),
+                'header' => __('Type'),
+                'index' => 'import',
+                'type' => 'options',
+                'options' => self::getOptionArray0(),
             ]
         );
 
@@ -109,7 +72,7 @@ class Grid extends Extended
             'code',
             [
                 'header' => __('Code'),
-                'index'  => 'code',
+                'index' => 'code',
             ]
         );
 
@@ -117,8 +80,8 @@ class Grid extends Extended
             'entity_id',
             [
                 'header' => __('Magento Entity ID'),
-                'index'  => 'entity_id',
-                'type'   => 'int',
+                'index' => 'entity_id',
+                'type' => 'int',
             ]
         );
 
@@ -126,11 +89,11 @@ class Grid extends Extended
             'created_at',
             [
                 'header' => __('Created'),
-                'index'  => 'created_at',
-                'type'   => 'datetime',
+                'index' => 'created_at',
+                'type' => 'datetime',
             ]
         );
-        
+
         $this->addExportType($this->getUrl('akeneomanager/*/exportCsv', ['_current' => true]), __('CSV'));
         $this->addExportType($this->getUrl('akeneomanager/*/exportExcel', ['_current' => true]), __('Excel XML'));
 
@@ -142,11 +105,7 @@ class Grid extends Extended
         return parent::_prepareColumns();
     }
 
-    
-    /**
-     * @return $this
-     */
-    protected function _prepareMassaction()
+    protected function _prepareMassaction(): Grid
     {
         $this->setMassactionIdField('id');
         $this->getMassactionBlock()->setFormFieldName('akeneo');
@@ -160,7 +119,7 @@ class Grid extends Extended
             ]
         );
 
-        $statuses = $this->_status->getOptionArray();
+        $statuses = $this->status->getOptionArray();
 
         $this->getMassactionBlock()->addItem(
             'status',
@@ -179,48 +138,36 @@ class Grid extends Extended
             ]
         );
 
-
         return $this;
     }
 
-    /**
-     * @return string
-     */
-    public function getGridUrl()
+    public function getGridUrl(): string
     {
         return $this->getUrl('akeneomanager/*/index', ['_current' => true]);
     }
 
-    /**
-     * @param \JustBetter\AkeneoBundle\Model\Akeneo|\Magento\Framework\Object $row
-     * @return string
-     */
-    public function getRowUrl($row)
+    public function getRowUrl($row): string
     {
-        return $this->getUrl(
-            'akeneomanager/*/edit',
-            ['id' => $row->getId()]
+        return $this->getUrl('akeneomanager/*/edit', ['id' => $row->getId()]);
+    }
+
+    public static function getOptionArray0(): array
+    {
+        return [
+            'family' => 'family',
+            'attribute' => 'attribute',
+            'category' => 'category',
+            'product' => 'product',
+            'option' => 'option',
+        ];
+    }
+
+    public static function getValueArray0(): array
+    {
+        return array_map(
+            fn($k, $v) => ['value' => $k, 'label' => $v],
+            array_keys(self::getOptionArray0()),
+            self::getOptionArray0()
         );
-    }
-    
-    public static function getOptionArray0()
-    {
-        $data_array = array();
-        $data_array['family'] = 'family';
-        $data_array['attribute'] = 'attribute';
-        $data_array['category'] = 'category';
-        $data_array['product'] = 'product';
-        $data_array['option'] = 'option';
-
-        return ($data_array);
-    }
-
-    public static function getValueArray0()
-    {
-        $data_array=array();
-        foreach (GridOption::getOptionArray0() as $k=> $v) {
-            $data_array[]=array('value'=>$k,'label'=>$v);
-        }
-        return($data_array);
     }
 }
