@@ -2,98 +2,67 @@
 
 namespace JustBetter\AkeneoBundle\Controller\Adminhtml\akeneo;
 
+use JustBetter\AkeneoBundle\Model\AkeneoFactory;
 use Magento\Backend\App\Action;
+use Magento\Backend\Model\Session;
+use Magento\Backend\Model\View\Result\Page;
+use Magento\Backend\Model\View\Result\Redirect;
+use Magento\Framework\App\Action\HttpGetActionInterface;
 use Magento\Framework\Registry;
 use Magento\Framework\View\Result\PageFactory;
 
-class Edit extends Action
+class Edit extends Action implements HttpGetActionInterface
 {
-    /**
-     * Core registry
-     *
-     * @var Registry
-     */
-    protected $_coreRegistry = null;
-
-    /**
-     * @var PageFactory
-     */
-    protected $resultPageFactory;
-
-    /**
-     * @param Action\Context $context
-     * @param PageFactory $resultPageFactory
-     * @param Registry $registry
-     */
     public function __construct(
         Action\Context $context,
-        PageFactory $resultPageFactory,
-        Registry $registry
+        protected PageFactory $resultPageFactory,
+        protected Registry $coreRegistry,
+        protected AkeneoFactory $akeneoFactory,
+        protected Session $session
     ) {
-        $this->resultPageFactory = $resultPageFactory;
-        $this->_coreRegistry = $registry;
         parent::__construct($context);
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    protected function _isAllowed()
+    protected function _isAllowed(): bool
     {
         return true;
     }
 
-    /**
-     * Init actions
-     *
-     * @return \Magento\Backend\Model\View\Result\Page
-     */
-    protected function _initAction()
+    protected function _initAction(): Page
     {
-        // load layout, set active menu and breadcrumbs
-        /** @var \Magento\Backend\Model\View\Result\Page $resultPage */
+        /** @var Page $resultPage */
         $resultPage = $this->resultPageFactory->create();
         $resultPage->setActiveMenu('JustBetter_AkeneoBundle::Akeneo')
             ->addBreadcrumb(__('JustBetter AkeneoBundle'), __('JustBetter AkeneoBundle'))
             ->addBreadcrumb(__('Manage Item'), __('Manage Item'));
+
         return $resultPage;
     }
 
-    /**
-     * Edit Item
-     *
-     * @return \Magento\Backend\Model\View\Result\Page|\Magento\Backend\Model\View\Result\Redirect
-     * @SuppressWarnings(PHPMD.NPathComplexity)
-     */
-    public function execute()
+    public function execute(): Page|Redirect
     {
-        // 1. Get ID and create model
-        $id = $this->getRequest()->getParam('id');
-        $model = $this->_objectManager->create('JustBetter\AkeneoBundle\Model\Akeneo');
+        $id = (int)$this->getRequest()->getParam('id');
+        $model = $this->akeneoFactory->create();
 
-        // 2. Initial checking
         if ($id) {
-            $model->load($id);
+            $model->load($id); // @phpstan-ignore-line
             if (!$model->getId()) {
-                $this->messageManager->addError(__('This item no longer exists.'));
-                /** \Magento\Backend\Model\View\Result\Redirect $resultRedirect */
+                $this->messageManager->addErrorMessage(__('This item no longer exists.'));
+                /** @var Redirect $resultRedirect */
                 $resultRedirect = $this->resultRedirectFactory->create();
 
                 return $resultRedirect->setPath('*/*/');
             }
         }
 
-        // 3. Set entered data if was error when we do save
-        $data = $this->_objectManager->get('Magento\Backend\Model\Session')->getFormData(true);
+        $data = $this->session->getFormData(true);
         if (!empty($data)) {
             $model->setData($data);
         }
 
-        // 4. Register model to use later in blocks
-        $this->_coreRegistry->register('akeneo', $model);
+        $this->coreRegistry->register('akeneo', $model);
 
-        // 5. Build edit form
-        /** @var \Magento\Backend\Model\View\Result\Page $resultPage */
+        /** @var Page $resultPage */
         $resultPage = $this->_initAction();
         $resultPage->setActiveMenu('JustBetter_AkeneoBundle::Akeneo');
         $resultPage->addBreadcrumb(__('JustBetter'), __('JustBetter'));
@@ -102,7 +71,6 @@ class Edit extends Action
             $id ? __('Edit Item') : __('New Item')
         );
         $resultPage->getConfig()->getTitle()->prepend($id ? __('Edit Item') : __('New Item'));
-        //$resultPage->getConfig()->getTitle()->prepend($model->getId() ? $model->getTitle() : __('New Item'));
 
         return $resultPage;
     }
