@@ -3,44 +3,28 @@
 namespace JustBetter\AkeneoBundle\Controller\Adminhtml\akeneo;
 
 use Magento\Backend\App\Action;
-use Magento\Backend\App\Action\Context;
+use Magento\Backend\Model\View\Result\Forward;
 use Magento\Backend\Model\View\Result\ForwardFactory;
+use Magento\Framework\App\Action\HttpGetActionInterface;
 
-class NewAction extends Action
+class NewAction extends Action implements HttpGetActionInterface
 {
-    /**
-     * @var \Magento\Backend\Model\View\Result\ForwardFactory
-     */
-    protected $resultForwardFactory;
-
-    /**
-     * @param Context $context
-     * @param ForwardFactory $resultForwardFactory
-     */
     public function __construct(
-        Context $context,
-        ForwardFactory $resultForwardFactory
+        Action\Context $context,
+        protected ForwardFactory $resultForwardFactory
     ) {
         parent::__construct($context);
-        $this->resultForwardFactory = $resultForwardFactory;
     }
 
-    /**
-     * Forward to edit
-     *
-     * @return \Magento\Backend\Model\View\Result\Forward
-     */
-    public function execute()
+    public function execute(): Forward
     {
-        /** @var \Magento\Backend\Model\View\Result\Forward $resultForward */
+        /** @var Forward $resultForward */
         $resultForward = $this->resultForwardFactory->create();
+
         return $resultForward->forward('edit');
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    protected function _isAllowed()
+    protected function _isAllowed(): bool
     {
         return true;
     }
