@@ -2,61 +2,63 @@
 
 namespace JustBetter\AkeneoBundle\Controller\Adminhtml\akeneo;
 
+use JustBetter\AkeneoBundle\Model\AkeneoFactory;
 use Magento\Backend\App\Action;
-use Magento\Framework\App\Filesystem\DirectoryList;
+use Magento\Backend\Model\Session;
+use Magento\Backend\Model\View\Result\Redirect;
+use Magento\Framework\App\Action\HttpPostActionInterface;
 
-class Save extends \Magento\Backend\App\Action
+class Save extends Action implements HttpPostActionInterface
 {
-
-    /**
-     * @param Action\Context $context
-     */
-    public function __construct(Action\Context $context)
-    {
+    public function __construct(
+        Action\Context $context,
+        protected AkeneoFactory $akeneoFactory,
+        protected Session $session
+    ) {
         parent::__construct($context);
     }
 
-    /**
-     * Save action
-     *
-     * @return \Magento\Framework\Controller\ResultInterface
-     */
-    public function execute()
+    public function execute(): Redirect
     {
         $data = $this->getRequest()->getPostValue();
 
-        /** @var \Magento\Backend\Model\View\Result\Redirect $resultRedirect */
+        /** @var Redirect $resultRedirect */
         $resultRedirect = $this->resultRedirectFactory->create();
-        if ($data) {
-            $model = $this->_objectManager->create('JustBetter\AkeneoBundle\Model\Akeneo');
 
-            $id = $this->getRequest()->getParam('id');
+        if ($data) {
+            $model = $this->akeneoFactory->create();
+
+            $id = (int)$this->getRequest()->getParam('id');
             if ($id) {
-                $model->load($id);
+                $model->load($id); // @phpstan-ignore-line
                 $model->setCreatedAt(date('Y-m-d H:i:s'));
             }
-            
+
             $model->setData($data);
 
             try {
-                $model->save();
-                $this->messageManager->addSuccess(__('The Akeneo has been saved.'));
-                $this->_objectManager->get('Magento\Backend\Model\Session')->setFormData(false);
+                $model->save(); // @phpstan-ignore-line
+                $this->messageManager->addSuccessMessage(__('The Akeneo has been saved.'));
+                $this->session->setFormData(false);
+
                 if ($this->getRequest()->getParam('back')) {
                     return $resultRedirect->setPath('*/*/edit', ['id' => $model->getId(), '_current' => true]);
                 }
+
                 return $resultRedirect->setPath('*/*/');
             } catch (\Magento\Framework\Exception\LocalizedException $e) {
-                $this->messageManager->addError($e->getMessage());
+                $this->messageManager->addErrorMessage($e->getMessage());
             } catch (\RuntimeException $e) {
-                $this->messageManager->addError($e->getMessage());
+                $this->messageManager->addErrorMessage($e->getMessage());
             } catch (\Exception $e) {
-                $this->messageManager->addException($e, __('Something went wrong while saving the Akeneo.'));
+                $this->messageManager->addExceptionMessage($e, __('Something went wrong while saving the Akeneo.'));
             }
 
-            $this->_getSession()->setFormData($data);
-            return $resultRedirect->setPath('*/*/edit', ['id' => $this->getRequest()->getParam('id')]);
+            $this->session->setFormData($data);
+
+            return $resultRedirect->setPath('*/*/edit', ['id' => $id]);
         }
+
         return $resultRedirect->setPath('*/*/');
     }
 }
