@@ -3,7 +3,6 @@
 namespace JustBetter\AkeneoBundle\Job;
 
 use Akeneo\Connector\Helper\Authenticator;
-use Akeneo\Pim\ApiClient\AkeneoPimClientInterface;
 use Akeneo\Pim\ApiClient\Pagination\ResourceCursor;
 use Akeneo\Pim\ApiClient\Search\SearchBuilder;
 use Magento\Eav\Api\AttributeRepositoryInterface;
@@ -18,33 +17,28 @@ class ImportMetricUnits
     protected const ENABLED_CONFIG_KEY = 'enablemetricunits';
     protected const CHANNEL_CONFIG_KEY = 'metric_conversion_channel';
 
-    protected Authenticator $authenticator;
-    protected AttributeRepositoryInterface $attributeRepository;
-    protected ScopeConfigInterface $config;
-
     public function __construct(
-        Authenticator $authenticator,
-        AttributeRepositoryInterface $attributeRepository,
-        ScopeConfigInterface $config
+        protected Authenticator $authenticator,
+        protected AttributeRepositoryInterface $attributeRepository,
+        protected ScopeConfigInterface $config
     ) {
-        $this->authenticator = $authenticator;
-        $this->attributeRepository = $attributeRepository;
-        $this->config = $config;
     }
 
     public function execute(?OutputInterface $output = null): void
     {
-        if (! $this->authenticator->getAkeneoApiClient()) {
+        if (!$this->authenticator->getAkeneoApiClient()) {
             if ($output) {
                 $output->writeln('<error>Akeneo client not configured!</error>');
             }
+
             return;
         }
-        
+
         if (!$this->config->getValue(static::CONFIG_PREFIX . static::ENABLED_CONFIG_KEY)) {
             if ($output) {
                 $output->writeln('<error>Metrics not enabled!</error>');
             }
+
             return;
         }
 
@@ -63,6 +57,7 @@ class ImportMetricUnits
                 if ($output) {
                     $output->writeln("<error>Skipping $code because it does not exist in Magento</error>");
                 }
+
                 continue;
             }
 
