@@ -2,19 +2,15 @@
 
 namespace JustBetter\AkeneoBundle\Model\ResourceModel\Akeneo;
 
+use JustBetter\AkeneoBundle\Model\Akeneo;
+use JustBetter\AkeneoBundle\Model\ResourceModel\Akeneo as AkeneoResourceModel;
 use Magento\Framework\Model\ResourceModel\Db\Collection\AbstractCollection;
 
 class Collection extends AbstractCollection
 {
-    /**
-     * Define resource model
-     *
-     * @return void
-     */
-    protected function _construct()
+    protected function _construct(): void
     {
-        $this->_init('JustBetter\AkeneoBundle\Model\Akeneo', 'JustBetter\AkeneoBundle\Model\ResourceModel\Akeneo');
+        $this->_init(Akeneo::class, AkeneoResourceModel::class);
         $this->_map['fields']['page_id'] = 'main_table.page_id';
     }
-
 }
