@@ -3,14 +3,14 @@
 namespace JustBetter\AkeneoBundle\Observer;
 
 use Akeneo\Connector\Executor\JobExecutor;
-use Magento\Framework\Event\ManagerInterface as EventManager;
+use Magento\Framework\Event\Manager;
 use Magento\Framework\Event\Observer;
 use Magento\Framework\Event\ObserverInterface;
 
 class ImportFinished implements ObserverInterface
 {
     public function __construct(
-        protected EventManager $eventManager
+        protected Manager $eventManager
     ) {
     }
 
@@ -21,10 +21,12 @@ class ImportFinished implements ObserverInterface
 
         $method = $executor->getMethod();
 
-        // Dispatch custom event before cleanCache step (last step before completion)
+        // Only dispatch for cleanCache method (end of import)
         if ($method === 'cleanCache') {
             $code = $executor->getCurrentJob()->getCode();
-            $this->eventManager->dispatch('akeneo_connector_import_finish_' . $code);
+            $event = 'akeneo_connector_import_finish_' . $code;
+
+            $this->eventManager->dispatch($event, ['import' => $executor]);
         }
     }
 }
