@@ -29,7 +29,7 @@ class CategoryExist
         foreach ($stores as $local => $affected) {
             foreach ($affected as $store) {
                 $columnName = 'url_key-' . $store['lang'];
-                
+
                 $query = "
                     UPDATE tmp_akeneo_connector_entities_category temp
                     LEFT JOIN catalog_category_entity_varchar eav ON (
@@ -43,14 +43,14 @@ class CategoryExist
                                 FROM eav_entity_type 
                                 WHERE entity_type_code = 'catalog_category'
                             )
-                            AND eav.`value` != ''
+                            AND eav.value != ''
                         )
                         AND eav.store_id = {$store['store_id']}
                     )
                     SET temp.`{$columnName}` = eav.value
                     WHERE temp._is_new = 0 AND eav.value IS NOT NULL
                 ";
-                
+
                 $connection->query($query);
             }
         }
